requirements.txt
altgraph==0.17.4
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==6.0.0
Deprecated==1.2.14
etelemetry==0.3.1
filelock==3.13.1
fitz==0.0.1.dev2
httplib2==0.22.0
idna==3.6
imageio==2.34.0
img2pdf==0.5.1
isodate==0.6.1
looseversion==1.3.0
lxml==5.1.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pefile==2023.2.7
pikepdf==8.13.0
pillow==10.2.0
prov==2.0.0
pydot==2.0.0
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
pywin32-ctypes==0.2.2
pyxnat==1.6.2
qcore==1.1.4
rdflib==7.0.0
requests==2.31.0
scipy==1.12.0
simplejson==3.19.2
six==1.16.0
traits==6.3.2
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
